+++ src/components/molecules/index.ts
@@ -18,4 +18,3 @@ export { default as ReviewItem } from "./ReviewItem/ReviewItem";
 export { default as SearchBar } from "./SearchBar/SearchBar";
 export { default as SidebarMenu } from "./SidebarMenu/SidebarMenu";
 export { default as UserMenu } from "./UserMenu/UserMenu";
-export { default as WishList } from "./WishList/WishList";
